+++ docs/CHANGELOG.md
@@ -15,6 +15,7 @@ Released on: ???
 
 - Excluded Vanquisher from being tracked in Sc/h tracker and Barn fishing timer.
 - Fixed Squid alert not working when Skyhanni's pet rarity chat setting enabled.
+- Fixed (I hope) the issue when own Dwarven Lantern / Umberella was not shown in the overlay sometimes after placing.
 
 ## Special thanks